refactor to better prepare for #18 and #3
